refactor: Split random
