+++ src/command.rs
@@ -1,13 +1,10 @@
 use {
     crate::{
         db::{FindOptions, MeigenDatabase},
-        model::Meigen,
         util::IteratorEditExt,
         Synced,
     },
     anyhow::{Context, Result},
-    rand::{rngs::StdRng, Rng, SeedableRng},
-    std::{future::Future, pin::Pin},
 };
 
 const MEIGEN_LENGTH_LIMIT: usize = 300;
@@ -79,44 +76,8 @@ pub use help::help;
 mod status;
 pub use status::status;
 
-pub async fn random(db: Synced<impl MeigenDatabase>, count: Option<u8>) -> Result<String> {
-    let (count, clamp_msg) = option!({
-        value: count,
-        default: 1,
-        min: 1,
-        max: 5,
-    });
-
-    fn get_random<'a>(
-        db: &'a Synced<impl MeigenDatabase>,
-        max: u32,
-    ) -> Pin<Box<dyn Future<Output = Result<Meigen>> + Send + 'a>> {
-        Box::pin(async move {
-            let pos = StdRng::from_rng(&mut rand::thread_rng())
-                .unwrap()
-                .gen_range(1..=max);
-
-            match db.read().await.load(pos).await? {
-                Some(e) => Ok(e),
-                None => get_random(db, max).await,
-            }
-        })
-    }
-
-    let mut meigens = Vec::with_capacity(count as _);
-    let max = db.read().await.get_current_id().await?;
-
-    for _ in 0..count {
-        meigens.push(get_random(&db, max).await?);
-    }
-
-    meigens.sort_by_key(|x| x.id);
-
-    let mut msg = meigens.into_iter().fold_list();
-    msg.insert_str(0, clamp_msg);
-
-    Ok(msg)
-}
+mod random;
+pub use random::random;
 
 pub async fn make(db: Synced<impl MeigenDatabase>, author: &str, content: &str) -> Result<String> {
     let strip = |s: &str| s.replace("`", "");
